+++ lib/domain/repository/settings_repository.dart
@@ -1,10 +1,10 @@
 abstract class SettingsRepository {
   Future<bool> getSfxEnabled();
   Future<bool> setSfxEnabled(bool isActivated);
-  Future<int> getSfxVolume();
-  Future<bool> setSfxVolume(int volume);
+  Future<double> getSfxVolume();
+  Future<bool> setSfxVolume(double volume);
   Future<bool> getMusicEnabled();
   Future<bool> setMusicEnabled(bool isActivated);
-  Future<int> getMusicVolume();
-  Future<bool> setMusicVolume(int volume);
+  Future<double> getMusicVolume();
+  Future<bool> setMusicVolume(double volume);
 }
